fix: shadow in the homepage and component split
