+++ src/components/TrustedBy/CardLabel.tsx
@@ -1,0 +1,3 @@
+export const CardLabel = ({ children }: { children: React.ReactNode }) => {
+  return <div className="text-base md:text-xl">{children}</div>
+}
